Refactor parser statements to make their names consistent with class names.
