+++ src/include/parser/insert_statement.h
@@ -13,8 +13,8 @@
 #pragma once
 
 #include "parser/sql_statement.h"
-#include "parser/statement_select.h"
 #include "optimizer/query_node_visitor.h"
+#include "select_statement.h"
 
 namespace peloton {
 namespace parser {
