+++ src/include/parser/statements.h
@@ -14,13 +14,13 @@
 
 // This is just for convenience
 
-#include "parser/statement_select.h"
-#include "parser/statement_create.h"
-#include "parser/statement_insert.h"
-#include "parser/statement_delete.h"
-#include "parser/statement_drop.h"
-#include "parser/statement_prepare.h"
-#include "parser/statement_execute.h"
-#include "parser/statement_transaction.h"
-#include "parser/statement_update.h"
-#include "parser/statement_copy.h"
+#include "copy_statement.h"
+#include "create_statement.h"
+#include "delete_statement.h"
+#include "drop_statement.h"
+#include "execute_statement.h"
+#include "insert_statement.h"
+#include "prepare_statement.h"
+#include "select_statement.h"
+#include "transaction_statement.h"
+#include "update_statement.h"
