+++ src/planner/create_plan.cpp
@@ -12,8 +12,8 @@
 
 #include "planner/create_plan.h"
 
+#include "../include/parser/create_statement.h"
 #include "storage/data_table.h"
-#include "parser/statement_create.h"
 #include "catalog/schema.h"
 #include "catalog/column.h"
